chore(inputs.redfish): Move thermal and power subsystems to their own files
